more modularization, some cleanup
