+++ lib/compilers.js
@@ -2,13 +2,13 @@
   var CoffeeScript, compilers, getUrlPath, jsExts, path, stripExt, _,
     __indexOf = [].indexOf || function(item) { for (var i = 0, l = this.length; i < l; i++) { if (i in this && this[i] === item) return i; } return -1; };
 
-  CoffeeScript = require('coffee-script');
-
   _ = require('underscore');
 
-  getUrlPath = require('./util').getUrlPath;
-
   path = require('path');
+
+  CoffeeScript = require('coffee-script');
+
+  getUrlPath = require('./util').getUrlPath;
 
   jsExts = function() {
     var ext;
