+++ lib/snockets.js
@@ -1,12 +1,10 @@
 (function() {
-  var DIRECTIVE, DepGraph, EXPLICIT_PATH, HEADER, HoldingQueue, Snockets, SourceMap, compilers, fs, getUrlPath, jsExts, minify, parseDirectives, path, sourceMapCat, stripExt, timeEq, _, _ref, _ref1,
+  var DIRECTIVE, DepGraph, EXPLICIT_PATH, HEADER, HoldingQueue, Snockets, compilers, fs, getUrlPath, jsExts, minify, parseDirectives, path, sourceMapCat, stripExt, timeEq, _, _ref, _ref1,
     __bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; },
     __slice = [].slice,
     __indexOf = [].indexOf || function(item) { for (var i = 0, l = this.length; i < l; i++) { if (i in this && this[i] === item) return i; } return -1; };
 
   DepGraph = require('dep-graph');
-
-  SourceMap = require('source-map');
 
   fs = require('fs');
 
@@ -14,13 +12,13 @@
 
   _ = require('underscore');
 
-  minify = require('./minification').minify;
-
-  _ref = require('./util'), timeEq = _ref.timeEq, getUrlPath = _ref.getUrlPath;
+  _ref = require('./util'), parseDirectives = _ref.parseDirectives, HoldingQueue = _ref.HoldingQueue, timeEq = _ref.timeEq, getUrlPath = _ref.getUrlPath, sourceMapCat = _ref.sourceMapCat, EXPLICIT_PATH = _ref.EXPLICIT_PATH, DIRECTIVE = _ref.DIRECTIVE, HEADER = _ref.HEADER;
 
   _ref1 = require('./compilers'), compilers = _ref1.compilers, jsExts = _ref1.jsExts, stripExt = _ref1.stripExt;
 
-  module.exports = Snockets = (function() {
+  minify = require('./minification').minify;
+
+  Snockets = (function() {
     function Snockets(options) {
       var errorStr, _base, _base1, _base2, _base3, _base4, _base5, _ref2, _ref3, _ref4, _ref5, _ref6, _ref7;
 
@@ -689,113 +687,8 @@
 
   })();
 
+  module.exports = Snockets;
+
   module.exports.compilers = compilers;
-
-  EXPLICIT_PATH = /^\/|:/;
-
-  HEADER = /(?:(\#\#\#.*\#\#\#\n*)|(\/\/.*\n*)|(\#.*\n*))+/;
-
-  DIRECTIVE = /^[\W]*=\s*(\w+.*?)(\*\\\/)?$/gm;
-
-  HoldingQueue = (function() {
-    function HoldingQueue(_arg) {
-      this.task = _arg.task, this.onComplete = _arg.onComplete;
-      this.holdKeys = [];
-    }
-
-    HoldingQueue.prototype.waitFor = function(key) {
-      return this.holdKeys.push(key);
-    };
-
-    HoldingQueue.prototype.unwaitFor = function(key) {
-      return this.holdKeys = _.without(this.holdKeys, key);
-    };
-
-    HoldingQueue.prototype.perform = function() {
-      var args, key,
-        _this = this;
-
-      key = arguments[0], args = 2 <= arguments.length ? __slice.call(arguments, 1) : [];
-      return this.task.apply(this, __slice.call(args).concat([function() {
-        return _this.unwaitFor(key);
-      }]));
-    };
-
-    HoldingQueue.prototype.finalize = function() {
-      var h,
-        _this = this;
-
-      if (this.holdKeys.length === 0) {
-        return this.onComplete();
-      } else {
-        return h = setInterval((function() {
-          if (_this.holdKeys.length === 0) {
-            _this.onComplete();
-            return clearInterval(h);
-          }
-        }), 10);
-      }
-    };
-
-    return HoldingQueue;
-
-  })();
-
-  parseDirectives = function(code) {
-    var header, match, _results;
-
-    code = code.replace(/[\r\t ]+$/gm, '\n');
-    if (!(match = HEADER.exec(code))) {
-      return [];
-    }
-    header = match[0];
-    _results = [];
-    while (match = DIRECTIVE.exec(header)) {
-      _results.push(match[1]);
-    }
-    return _results;
-  };
-
-  sourceMapCat = function(opts) {
-    var combinedGeneratedLine, generated, original, originalLastLine, _i, _len, _original, _ref2;
-
-    generated = new SourceMap.SourceMapGenerator({
-      file: opts.filename
-    });
-    combinedGeneratedLine = 1;
-    _ref2 = opts.maps;
-    for (_i = 0, _len = _ref2.length; _i < _len; _i++) {
-      _original = _ref2[_i];
-      if ((_original.empty != null) && _original.empty === true) {
-        combinedGeneratedLine += _original.numLines;
-        continue;
-      }
-      original = new SourceMap.SourceMapConsumer(_original);
-      originalLastLine = null;
-      original.eachMapping(function(mapping) {
-        var e;
-
-        try {
-          generated.addMapping({
-            generated: {
-              line: combinedGeneratedLine + mapping.generatedLine,
-              column: mapping.generatedColumn
-            },
-            original: {
-              line: mapping.originalLine,
-              column: mapping.originalColumn
-            },
-            source: mapping.source
-          });
-        } catch (_error) {
-          e = _error;
-          throw new Error("Invalid Mapping: " + (JSON.stringify(mapping)));
-        }
-        return originalLastLine = mapping.generatedLine;
-      });
-      combinedGeneratedLine += originalLastLine;
-    }
-    return JSON.parse(generated.toString());
-  };
 
 }).call(this);
