+++ package.json
@@ -1,7 +1,7 @@
 {
   "author": {
     "name": "Trevor Burnham",
-    "url": "http://trevorburnham.com"
+    "url": "http://trevorburnham.com/"
   },
   "contributors": [
     {
